ai-service/pyproject.toml
[project]
name = "ai-interview-ai-service"
version = "0.1.0"
description = "AI 模拟面试系统的 FastAPI 内部 AI 服务"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "celery[redis]>=5.4.0,<6.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "pydantic-settings>=2.7.0,<3.0.0",
  "python-multipart>=0.0.18,<1.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
